MAINT: Make `typing._callback` available at runtime
